Move stuff from utils to parse
