+++ bot/submission/submit.py
@@ -20,6 +20,7 @@ from pydantic import ValidationError
 from typing_extensions import override
 
 from bot import utils
+from bot.submission.parse import remove_markdown, parse_build_title
 from bot.vote_session import AbstractVoteSession
 from bot.submission.ui import BuildSubmissionForm, ConfirmationView
 from database import message as msg
@@ -27,7 +28,7 @@ from database.builds import get_all_builds, Build
 from database import DatabaseManager
 from database.enums import Status, Category
 from bot._types import SubmissionCommandResponse, GuildMessageable
-from bot.utils import RunningMessage, parse_dimensions, parse_build_title, remove_markdown, is_owner_server
+from bot.utils import RunningMessage, parse_dimensions, is_owner_server
 from database.message import get_build_id_by_message
 from database.schema import TypeRecord
 from database.server_settings import get_server_setting
